+++ NEWS.md
@@ -1,5 +1,6 @@
 # mlr3tuning 0.12.1.9000
 
+* feat: Allow to pass `Tuner` objects as `method` in `tune()` and `auto_tuner()`.
 * docs: Link `Tuner` to help page of `bbotk::Optimizer`.
 * feat: `Tuner` objects have the optional field `$label` now.
 * feat: `as.data.table()` functions for objects of class `Dictionary` have been extended with additional columns.